Add files via upload
Please download the executable file, and install it on your computer. You can run it on Microsoft Windows.
The developed tool predicts the backbone curve of the endplate connections equipped with SMA bolts. It will be helpful to eliminate the need for a detailed finite element modeling of the connections.